Revert "deps: upgrade openssl to 1.0.2a"
This reverts commit e4872d7405c60da50692e3c4cd5a9df2e96a5680.

+++ deps/openssl/openssl/crypto/objects/obj_dat.c
@@ -142,7 +142,7 @@ static unsigned long added_obj_hash(const ADDED_OBJ *ca)
         return 0;
     }
     ret &= 0x3fffffffL;
-    ret |= ((unsigned long)ca->type) << 30L;
+    ret |= ca->type << 30L;
     return (ret);
 }
